start to consolidate env variables etc
