+++ backend/.env.example
@@ -2,3 +2,6 @@
 
 BACKEND_PORT=3000
 JWT_SECRET=my_secret_token
+
+RETOOL_API_KEY=retool_01gh1qd61vcg45x4zbvg8h4pzr
+RETOOL_URL=retool.shopco.partners
