+++ backend/config.js
@@ -1,3 +1,5 @@
+exports.retool_api_key = 'retool_01gh9pd7qre20ptp9b51zrpfg8'
+
 exports.default_user = {
     username: 'john.doe@example.com',
     roles: ['gold', 'silver', 'bronze']
