+++ backend/server.js
@@ -2,8 +2,8 @@ var express = require('express');
 var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
-var retoolRouter = require('./api/retool');
 var indexRouter = require('./routes/index');
+var retoolRouter = require('./routes/retool');
 
 var app = express();
